requirements.txt
python-telegram-bot[job-queue]==21.6
httpx==0.27.2
pytz==2024.1
google-generativeai==0.8.3
fastapi==0.115.5
uvicorn==0.32.1
psycopg2-binary==2.9.10
